Added role configuration & final updates.

+++ src/main/java/com/javaholics/lakesidehotel/service/UserServiceImpl.java
@@ -43,7 +43,10 @@ public class UserServiceImpl implements UserService {
     @Transactional
     @Override
     public void deleteUser(String email) {
-        userRepository.deleteByEmail(email);
+        User theUser = getUser(email);
+        if (theUser != null) {
+            userRepository.deleteByEmail(email);
+        }
     }
 
     @Override
